// web/src/lib/api/model/assetKind.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Silo
 * OpenAPI spec version: 0.1.0
 */

export type AssetKind = (typeof AssetKind)[keyof typeof AssetKind]

export const AssetKind = {
  cover: "cover",
  background: "background",
  landscape: "landscape",
  logo: "logo",
  icon: "icon",
  screenshot: "screenshot",
  video_thumbnail: "video_thumbnail",
} as const